requirements.txt
APScheduler==3.8.0
discord.py==2.1.0
emoji==1.5.0
Flask==2.2.2
humanfriendly==10.0
inflect==5.4.0
psycopg2==2.9.3
pytest==7.2.1
requests==2.25.1
